Add files via upload
This file should do what the lab data team wanted us to do! It uses Chris' csv extraction function and a repurposed version of my coincidence finding code. Once we've checked that it works, we should be able to send it to the lab data team for them to use right away :)